InversionCounting PerfTest

+++ Algorithms/InversionCounting/NLogN.cs
@@ -6,32 +6,32 @@ namespace Algorithms.InversionCounting
 {
 	public class NLogN
 	{
-		public int Count(int[] elements)
+		public long Count(int[] elements)
 		{
 			return SortAndCount(elements.ToList()).Item1;
 		}
 
-		public Tuple<int, List<int>> SortAndCount(List<int> list)
+		public Tuple<long, List<int>> SortAndCount(List<int> list)
 		{
 			if (list.Count <= 1)
-				return new Tuple<int, List<int>>(0, list);
+				return new Tuple<long, List<int>>(0, list);
 
 			int middle = list.Count / 2;
 			var leftList = list.GetRange(0, middle);
 			var rightList = list.GetRange(middle, list.Count - leftList.Count);
 
-			Tuple<int, List<int>> leftResult = SortAndCount(leftList);
-			Tuple<int, List<int>> rightResult = SortAndCount(rightList);
+			Tuple<long, List<int>> leftResult = SortAndCount(leftList);
+			Tuple<long, List<int>> rightResult = SortAndCount(rightList);
 
-			Tuple<int, List<int>> mergeResult = MergeAndCount(leftResult.Item2, rightResult.Item2);
+			Tuple<long, List<int>> mergeResult = MergeAndCount(leftResult.Item2, rightResult.Item2);
 
-			return new Tuple<int, List<int>>(leftResult.Item1 + rightResult.Item1 + mergeResult.Item1,
+			return new Tuple<long, List<int>>(leftResult.Item1 + rightResult.Item1 + mergeResult.Item1,
 			mergeResult.Item2);
 		}
 
-		private Tuple<int, List<int>> MergeAndCount(List<int> leftList, List<int> rightList)
+		private Tuple<long, List<int>> MergeAndCount(List<int> leftList, List<int> rightList)
 		{
-			int inversions = 0;
+			long inversions = 0;
 			var outputList = new List<int>();
 			int i = 0, j = 0;
 
@@ -55,7 +55,7 @@ namespace Algorithms.InversionCounting
 			else if (j < rightList.Count)
 				outputList.AddRange(rightList.GetRange(j, rightList.Count - j));
 
-			return new Tuple<int, List<int>>(inversions, outputList);
+			return new Tuple<long, List<int>>(inversions, outputList);
 		}
 	}
 }
